// frontend/src/api/generated/platform.ts
export type WebsiteProjectProfileRequest = {
  name: string
  domain: string
  country: string
  target_market: string
  language: string
  products?: Array<string>
  keywords?: Array<string>
  target_urls?: Array<string>
}